[API BREAKING] Merge ValidateObject into Object.
