+++ lib/xgit/plumbing/hash_object.ex
@@ -10,7 +10,6 @@ defmodule Xgit.Plumbing.HashObject do
   alias Xgit.Core.Object
   alias Xgit.Core.ObjectId
   alias Xgit.Core.ObjectType
-  alias Xgit.Core.ValidateObject
   alias Xgit.Repository
 
   @doc ~S"""
@@ -54,12 +53,12 @@ defmodule Xgit.Plumbing.HashObject do
 
   * `Xgit.Core.FilePath.check_path/2`
   * `Xgit.Core.FilePath.check_path_segment/2`
-  * `Xgit.Core.ValidateObject.check/2`
+  * `Xgit.Core.Object.check/2`
   * `Xgit.Repository.put_loose_object/2`.
   """
   @spec run(content :: ContentSource.t(), type: ObjectType.t() | nil) ::
           {:ok, ObjectId.t()}
-          | {:error, reason :: ValidateObject.check_reason()}
+          | {:error, reason :: Object.check_reason()}
           | {:error, reason :: FilePath.check_path_reason()}
           | {:error, reason :: FilePath.check_path_segment_reason()}
           | {:error, reason :: Repository.put_loose_object_reason()}
@@ -128,7 +127,7 @@ defmodule Xgit.Plumbing.HashObject do
   defp validate_content(object, false = _validate?), do: {:ok, object}
 
   defp validate_content(%Object{content: content} = object, _validate?) when is_list(content) do
-    case ValidateObject.check(object) do
+    case Object.check(object) do
       :ok -> {:ok, object}
       {:error, reason} -> {:error, reason}
     end
